alx_travel_app_0x02/requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
